백업/페이즈1(개발중) 백업/requirements.txt
requests
google-genai
groq
slack_sdk
PyYAML
yara-python
docker
supabase
pytz
tenacity